# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for the Dermify application",
        "title": "Dermify API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/outcome": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the clinical outcome for a session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outcome"
                ],
                "summary": "Get outcome",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionOutcomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the clinical outcome for a session. Requires version for optimistic locking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outcome"
                ],
                "summary": "Update outcome",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated outcome details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateOutcomeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionOutcomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records the clinical outcome for a session (one per session).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outcome"
                ],
                "summary": "Record outcome",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Outcome details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.recordOutcomeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.SessionOutcomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH_INVALID_CREDENTIALS"
                },
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
...
        "handlers.SessionOutcomeResponse": {
            "type": "object",
            "properties": {
                "aftercare_notes": {
                    "type": "string"
                },
                "contact_info": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "endpoint_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_up_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "type": "string"
                },
                "outcome_status": {
                    "type": "string",
                    "example": "completed"
                },
                "red_flags_text": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.recordOutcomeRequest": {
            "type": "object",
            "properties": {
                "aftercare_notes": {
                    "type": "string"
                },
                "contact_info": {
                    "type": "string"
                },
                "endpoint_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_up_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "outcome_status": {
                    "type": "string"
                },
                "red_flags_text": {
                    "type": "string"
                }
            }
        },
...
        "handlers.updateOutcomeRequest": {
            "type": "object",
            "properties": {
                "aftercare_notes": {
                    "type": "string"
                },
                "contact_info": {
                    "type": "string"
                },
                "endpoint_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "follow_up_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "outcome_status": {
                    "type": "string"
                },
                "red_flags_text": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}